Chrome browser reveals database locking issue
I have so far avoided to lock the database during upload
session generation etc. Now it turns out that it ___has___
to . be . locked though, at any cost, because otherwise it can't be
ensured, that only one analysis is created per upload session. I.e.
even thought committing the uploadsession object will faile because the
other request has changed it, creation of the Analysis object succeeds,
which is really bad... :-(

The fix is ugly, but it will be easy to fix properly later on.
For now it works.